refactor: Added test fixtures and refactored PublisherTest to use them
