+++ block-node/app/build.gradle.kts
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 plugins {
     id("org.hiero.gradle.module.library")
+    id("org.hiero.gradle.feature.test-fixtures")
     id("application")
 }
